Fix: pricingBounds validations
