+++ src/contexts/CreateCampaignFormContext/CreateCampaignFormContext.tsx
@@ -1,10 +1,13 @@
 import { Button } from '@mantine/core'
 import { createFormContext } from '@mantine/form'
-import { validateCreateCampaignFrom } from 'helpers/validators'
+import { parseBigNumTokenAmountToDecimal } from 'helpers'
 import useAccount from 'hooks/useAccount'
 import useCreateCampaignContext from 'hooks/useCreateCampaignContext'
 import { FC, PropsWithChildren, useEffect } from 'react'
 import { CampaignUI } from 'types'
+
+const MIN_CAMPAIGN_BUDGET_VALUE = 500
+const MIN_CPM_VALUE = 0.1
 
 const [CrCampaignFormProvider, useCreateCampaignFormContext, useCreateCampaignForm] =
   createFormContext<CampaignUI>()
@@ -15,9 +18,91 @@ const CreateCampaignFormProvider: FC<PropsWithChildren> = ({ children }) => {
     adexAccount: { availableBalance, balanceToken }
   } = useAccount()
 
+  const validateBudget = (value: string) => {
+    const formattedToken = Number(
+      parseBigNumTokenAmountToDecimal(availableBalance, balanceToken.decimals)
+    )
+    return formattedToken < Number(value)
+  }
+
+  const validateCPMMin = (formValues: CampaignUI) => {
+    const {
+      cpmPricingBounds: { min, max }
+    } = formValues
+
+    if (Number(min) === 0 || Number.isNaN(Number(min))) {
+      return 'Enter CPM min value or a valid number'
+    }
+
+    if (Number(min) < MIN_CPM_VALUE) {
+      return `CPM min can not be lower than ${MIN_CPM_VALUE}`
+    }
+    if (min >= max) {
+      return 'CPM min can not be greater than CPM max'
+    }
+    return null
+  }
+
+  const validateCPMMax = (formValues: CampaignUI) => {
+    const {
+      cpmPricingBounds: { min, max }
+    } = formValues
+
+    if (Number(max) === 0 || Number.isNaN(Number(max))) {
+      return 'Enter CPM max value or a valid number'
+    }
+    if (Number(max) <= 0) {
+      return 'CPM max should be greater than 0'
+    }
+    if (max <= min) {
+      return 'CPM max can not be lower than CPM min'
+    }
+    return null
+  }
+
   const form = useCreateCampaignForm({
     initialValues: campaign,
-    validate: validateCreateCampaignFrom(availableBalance, balanceToken)
+    validate: {
+      paymentModel: (value: string) => {
+        if (value === '') return 'Select payment method'
+        return null
+      },
+      currency: (value: string) => {
+        if (value === '') return 'Select currency'
+        return null
+      },
+      campaignBudget: (value: any) => {
+        if (value === '' || Number.isNaN(Number(value))) {
+          return 'Enter campaign budget or a valid number'
+        }
+
+        if (parseFloat(value) < MIN_CAMPAIGN_BUDGET_VALUE) {
+          return `Campaign budget can not be lower than ${MIN_CAMPAIGN_BUDGET_VALUE}`
+        }
+
+        if (validateBudget(value)) {
+          return 'Available balance is lower than the campaign budget'
+        }
+        return null
+      },
+      cpmPricingBounds: {
+        min: () => {
+          return validateCPMMin(form.values)
+        },
+        max: () => {
+          return validateCPMMax(form.values)
+        }
+      },
+      title: (value: string) => {
+        if (value === '') {
+          return 'Enter Title'
+        }
+        if (value.length < 2) {
+          return 'Campaign name must have at least 2 letters'
+        }
+        return null
+      }
+    }
   })
 
   const handleSubmit = (event: any) => {
