+++ src/helpers/validators.ts
@@ -1,71 +1,3 @@
-import { Token } from 'types'
-import { parseBigNumTokenAmountToDecimal } from './balances'
-
-const MIN_CAMPAIGN_BUDGET_VALUE = 500
-const MIN_CPM_VALUE = 0.1
-
-export const validateCreateCampaignFrom = (availableBalance: bigint, balanceToken: Token) => {
-  const validateBudget = (value: string) => {
-    const formattedToken = Number(
-      parseBigNumTokenAmountToDecimal(availableBalance, balanceToken.decimals)
-    )
-    return formattedToken < Number(value)
-  }
-
-  return {
-    paymentModel: (value: string) => {
-      if (value === '') return 'Select payment method'
-      return null
-    },
-    currency: (value: string) => {
-      if (value === '') return 'Select currency'
-      return null
-    },
-    campaignBudget: (value: any) => {
-      if (value === '' || Number.isNaN(Number(value))) {
-        return 'Enter campaign budget or a valid number'
-      }
-
-      if (parseFloat(value) < MIN_CAMPAIGN_BUDGET_VALUE) {
-        return `Campaign budget can not be lower than ${MIN_CAMPAIGN_BUDGET_VALUE}`
-      }
-
-      if (validateBudget(value)) {
-        return 'Available balance is lower than the campaign budget'
-      }
-      return null
-    },
-    'cpmPricingBounds.min': (value: string) => {
-      if (value === '' || Number.isNaN(Number(value))) {
-        return 'Enter CPM min value or a valid number'
-      }
-
-      if (parseFloat(value) < MIN_CPM_VALUE) {
-        return `CPM min can not be lower than ${MIN_CPM_VALUE}`
-      }
-      return null
-    },
-    'cpmPricingBounds.max': (value: string) => {
-      if (value === '' || Number.isNaN(Number(value))) {
-        return 'Enter CPM max value or a valid number'
-      }
-      if (parseFloat(value) <= 0) {
-        return 'CPM max should be greater than 0'
-      }
-      return null
-    },
-    title: (value: string) => {
-      if (value === '') {
-        return 'Enter Title'
-      }
-      if (value.length < 2) {
-        return 'Campaign name must have at least 2 letters'
-      }
-      return null
-    }
-  }
-}
-
 export const isValidHttpUrl = (inputURL: string) => {
   let url
